Ajout de la suppression quand on crée une vente

+++ PopupStore/PricePage.xaml.cs
@@ -65,12 +65,15 @@ namespace PopupStore
         {
             try
             {
-                int id = ((DAL.DB.Price)priceDataGrid.SelectedItem).Id;
-                DAL.DB.Price price = BU.PriceService.GetPrice(id);
-                if (price.Products.Count > 0)
-                    throw new Exception("Le prix est associé à au moins un produit !");
-                BU.PriceService.DeletePrice(price);
-                InitDataContext();
+                if ((DAL.DB.Price)priceDataGrid.SelectedItem != null)
+                {
+                    int id = ((DAL.DB.Price)priceDataGrid.SelectedItem).Id;
+                    DAL.DB.Price price = BU.PriceService.GetPrice(id);
+                    if (price.Products.Count > 0)
+                        throw new Exception("Le prix est associé à au moins un produit !");
+                    BU.PriceService.DeletePrice(price);
+                    InitDataContext();
+                }
             }
             catch(Exception exception)
             {
